+++ CHANGES.rst
@@ -15,6 +15,7 @@ Fixed
 - Password reset no longer bypasses MFA.
 - Usernames don't appear in the photos URLs.
 - Include missing Babel data in the pyinstaller binary. :issue:`318`
+- OIDC discovery signing algorithm claims are now dynamically computed from server keys. :issue:`323`
 
 [0.1.0] - 2025-11-13
 --------------------